bot/requirements.txt
SQLAlchemy==2.0.31
fastapi==0.116.1
uvicorn==0.30.1
bcrypt==4.0.1
dataclasses==0.6
sentencepiece==0.2.0
pydantic==2.7.1
pydantic-settings==2.1.0
asyncpg==0.29.0
torch==2.5.1
transformers==4.40.1
email-validator==2.2.0
python-telegram-bot==20.3
requests

